pyproject.toml
[project]
name = "python-ws"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "googleapis-common-protos>=1.72.0",
    "grpcio>=1.76.0",
    "protobuf==6.33.1",
]

[dependency-groups]
dev = [
    "grpcio-tools>=1.76.0",
]
